Add helper output for connecting after running add-to-ssh-agent
The function definition header has help how to connect
to the server whose key we've added to the ssh-agent, on
the other hand the tabular output is both different order
and would still require knowledge of the ssh parameters.

This change fills in the command with the server variables
and adds some extra flags that help to be less noisy later:
not adding the server's key to the known hosts file, and
not checking the server's key.

```
Hostname          Port  Username
123.123.123.123   12345  faculty

Connect to the server by:
ssh faculty@123.123.123.123 -p 12345 -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no
```

It should work both on Linux/Mac and Windows as the relevant
null device is substituted in the command example.

Signed-off-by: Gergely Imreh <gergely.imreh@faculty.ai>