Rename dataset to datasets
In #23, I asked @liamcoatman to make the name of the new datasets
command group singular, as everything else in the CLI is singular.
However, on merging #23, I tested out the CLI for myself and on
reflection this feels very unnatural as the corresponding feature in the
platform is always referred to in the singular.

This PR proposes to revert to the plural naming.